requirements.txt
# AI and computer vision
ultralytics>=8.3.0
opencv-python>=4.8.0
supervision>=0.16.0

# Hardware and IO
pyserial>=3.5
obd>=0.7.1

# Data processing
numpy>=1.24.0

# API and networking
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
python-multipart>=0.0.6
jinja2>=3.1.0
httpx>=0.27.0
